requirements.txt
slack_bolt==1.18.1
slack_sdk==3.27.1
google-api-python-client==2.133.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
cachetools==5.3.3
